extract preview image
